requirements.txt
fastmcp>=0.1.2
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
kafka-python>=2.0.2
pydantic>=2.7.1
python-dotenv>=1.0.1
